+++ esp-wifi/src/timer/riscv.rs
@@ -1,6 +1,10 @@
 use core::cell::RefCell;
 
 use critical_section::Mutex;
+use esp_hal::{
+    interrupt::InterruptHandler,
+    timer::{ErasedTimer, PeriodicTimer},
+};
 #[cfg(any(feature = "esp32c6", feature = "esp32h2"))]
 use peripherals::INTPRI as SystemPeripheral;
 #[cfg(not(any(feature = "esp32c6", feature = "esp32h2")))]
@@ -10,46 +14,32 @@ use crate::{
     hal::{
         interrupt::{self, TrapFrame},
         peripherals::{self, Interrupt},
-        prelude::*,
         riscv,
-        timer::systimer::{Alarm, Periodic, Target},
     },
     preempt::preempt::task_switch,
 };
 
 /// The timer responsible for time slicing.
-pub type TimeBase = Alarm<Target, esp_hal::Blocking, 0>;
-static ALARM0: Mutex<RefCell<Option<Alarm<Periodic, esp_hal::Blocking, 0>>>> =
-    Mutex::new(RefCell::new(None));
-const TIMESLICE_FREQUENCY: fugit::HertzU32 = fugit::HertzU32::from_raw(crate::CONFIG.tick_rate_hz);
+pub type TimeBase = PeriodicTimer<ErasedTimer>;
+static ALARM0: Mutex<RefCell<Option<TimeBase>>> = Mutex::new(RefCell::new(None));
+const TIMESLICE_FREQUENCY: fugit::HertzU64 =
+    fugit::HertzU64::from_raw(crate::CONFIG.tick_rate_hz as u64);
 
 // Time keeping
 pub const TICKS_PER_SECOND: u64 = 1_000_000;
 
-pub fn setup_timer(systimer: TimeBase) -> Result<(), esp_hal::timer::Error> {
+pub fn setup_timer(mut alarm0: TimeBase) -> Result<(), esp_hal::timer::Error> {
     // make sure the scheduling won't start before everything is setup
     riscv::interrupt::disable();
 
-    let alarm0 = systimer.into_periodic();
-    alarm0.set_period(TIMESLICE_FREQUENCY.into_duration());
-    alarm0.clear_interrupt();
-    alarm0.enable_interrupt(true);
+    let cb: extern "C" fn() = unsafe { core::mem::transmute(handler as *const ()) };
+    alarm0.set_interrupt_handler(InterruptHandler::new(cb, interrupt::Priority::Priority1));
+    alarm0.start(TIMESLICE_FREQUENCY.into_duration())?;
+    critical_section::with(|cs| {
+        alarm0.enable_interrupt(true);
+        ALARM0.borrow_ref_mut(cs).replace(alarm0);
+    });
 
-    critical_section::with(|cs| ALARM0.borrow_ref_mut(cs).replace(alarm0));
-
-    unsafe {
-        interrupt::bind_interrupt(
-            Interrupt::SYSTIMER_TARGET0,
-            core::mem::transmute::<*const (), unsafe extern "C" fn()>(
-                systimer_target0 as *const (),
-            ),
-        );
-    }
-
-    unwrap!(interrupt::enable(
-        Interrupt::SYSTIMER_TARGET0,
-        interrupt::Priority::Priority1,
-    ));
     Ok(())
 }
 
@@ -64,7 +54,7 @@ pub fn setup_multitasking() {
     }
 }
 
-extern "C" fn systimer_target0(trap_frame: &mut TrapFrame) {
+extern "C" fn handler(trap_frame: &mut TrapFrame) {
     // clear the systimer intr
     critical_section::with(|cs| {
         unwrap!(ALARM0.borrow_ref_mut(cs).as_mut()).clear_interrupt();
@@ -83,10 +73,8 @@ extern "C" fn FROM_CPU_INTR3(trap_frame: &mut TrapFrame) {
     }
 
     critical_section::with(|cs| {
-        let alarm0 = ALARM0.borrow_ref(cs);
-        let alarm0 = unwrap!(alarm0.as_ref());
-
-        alarm0.set_period(TIMESLICE_FREQUENCY.into_duration());
+        let mut alarm0 = ALARM0.borrow_ref_mut(cs);
+        let alarm0 = unwrap!(alarm0.as_mut());
         alarm0.clear_interrupt();
     });