+++ examples/src/bin/embassy_multiprio.rs
@@ -15,7 +15,7 @@
 // The interrupt-executor is created in `main` and is used to spawn `high_prio`.
 
 //% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: embassy embassy-time-timg0 embassy-generic-timers
+//% FEATURES: embassy esp-hal-embassy/log esp-hal-embassy/integrated-timers
 
 #![no_std]
 #![no_main]
@@ -29,11 +29,21 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::timg::TimerGroup,
+    timer::{/*systimer::SystemTimer,*/ timg::TimerGroup, ErasedTimer, OneShotTimer},
 };
 use esp_hal_embassy::InterruptExecutor;
 use esp_println::println;
 use static_cell::StaticCell;
+
+// When you are okay with using a nightly compiler it's better to use https://docs.rs/static_cell/2.1.0/static_cell/macro.make_static.html
+macro_rules! mk_static {
+    ($t:ty,$val:expr) => {{
+        static STATIC_CELL: static_cell::StaticCell<$t> = static_cell::StaticCell::new();
+        #[deny(unused_attributes)]
+        let x = STATIC_CELL.uninit().write(($val));
+        x
+    }};
+}
 
 /// Periodically print something.
 #[embassy_executor::task]
@@ -78,8 +88,21 @@ async fn main(low_prio_spawner: Spawner) {
     let system = SystemControl::new(peripherals.SYSTEM);
     let clocks = ClockControl::boot_defaults(system.clock_control).freeze();
 
-    let timg0 = TimerGroup::new_async(peripherals.TIMG0, &clocks);
-    esp_hal_embassy::init(&clocks, timg0);
+    let timg0 = TimerGroup::new(peripherals.TIMG0, &clocks, None);
+    let timer0 = OneShotTimer::new(timg0.timer0.into());
+    #[cfg(not(feature = "esp32c2"))]
+    let timer1 = {
+        let timg1 = TimerGroup::new(peripherals.TIMG1, &clocks, None);
+        OneShotTimer::new(timg1.timer0.into())
+    };
+    #[cfg(feature = "esp32c2")]
+    let timer1 = {
+        let systimer = esp_hal::timer::systimer::SystemTimer::new(peripherals.SYSTIMER);
+        OneShotTimer::new(systimer.alarm0.into())
+    };
+    let timers = [timer0, timer1];
+    let timers = mk_static!([OneShotTimer<ErasedTimer>; 2], timers);
+    esp_hal_embassy::init(&clocks, timers);
 
     static EXECUTOR: StaticCell<InterruptExecutor<2>> = StaticCell::new();
     let executor = InterruptExecutor::new(system.software_interrupt_control.software_interrupt2);